+++ nimare/base/decode.py
@@ -1,11 +1,13 @@
 """
 Base classes for decoding/encoding.
 """
+from .base import Estimator
 
 
-class Decoder(object):
+class Decoder(Estimator):
     """
-    Base class for decoding.
+    Base class for decoders. Decoders act as Estimator, in that they take in
+    Datasets and return Results.
     """
     def __init__(self, dataset):
         pass
